Add the other static values to `DifficultyAttributes` and change the API to include the beatmap
